requirements.txt
python-dotenv
requests
whisper
TTS
PySide6
yt_dlp
python-docx
